+++ CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixed:
-- Chart header cut off
-
 ### Changed:
 - Charts hidden in journal cards
+- New bottom navigation
+
+### Fixed:
+- Chart header cut off
 
 ## [0.8.155] - 2022-09-25
 ### Changed: